Fix console.error left in builtins
